rewrite a bugfix in throttle signalling
state_changed is used to tell the throttler that
work has been done.

Though, the code does a OR to save that state across a buycheck
and a sellcheck. But the IF statement needs to get a clean
result from the buy/sell check, not a ORed one. Else the sell
will always go through, if a buy has been executed above.

To test this we need a testcase that sets up two trades,
but to do this we need to mock another exchange function.

Instead of mocking (eventually full mocking) the exchange,
introduce a testdummy-exchange, that is used for dry-run and testing.